Checking admin
Auth::user(): This retrieves the currently authenticated user.Auth::user()-&gt;utype === 'ADM': This checks if the authenticated user's type (utype) is 'ADM' (indicating an Admin user).If the user is an Admin:return $next($request): The request is allowed to proceed to the next middleware or the intended controller/action. This means that Admin users are granted access to the route this middleware is protecting.If the user is not an Admin:session()-&gt;flush(): This clears all session data, effectively logging out the user by removing their session.return redirect()-&gt;route('login'): This redirects the user to the login page. This means that non-Admin users are not granted access and are instead redirected to log in again.